LPS-111959 Migrates LayoutSelector to use React only and applies some naming standards
