+++ modules/apps/layout/layout-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/LayoutSelector.js
@@ -12,173 +12,133 @@
  * details.
  */
 
-import 'dynamic-data-mapping-form-field-type';
-import Component from 'metal-component';
-import Soy from 'metal-soy';
-import {Config} from 'metal-state';
+import ClayButton from '@clayui/button';
+import ClayForm, {ClayInput} from '@clayui/form';
+import {
+	FieldBaseProxy,
+	connectStore,
+	getConnectedReactComponentAdapter,
+} from 'dynamic-data-mapping-form-field-type';
+import {ItemSelectorDialog} from 'frontend-js-web';
+import React, {useEffect, useState} from 'react';
 
-import './LayoutSelectorAdapter.soy';
+const LayoutSelector = ({
+	disabled,
+	inputValue,
+	itemSelectorURL,
+	name,
+	onChange,
+	portletNamespace,
+}) => {
+	const [layout, setLayout] = useState({});
 
-import './LayoutSelectorRegister.soy';
-
-import './ReactLayoutSelectorAdapter';
-import templates from './LayoutSelector.soy';
-
-class LayoutSelector extends Component {
-	dispatchEvent(event, name, value) {
-		this.emit(name, {
-			fieldInstance: this,
-			originalEvent: event,
-			value,
+	useEffect(() => {
+		setLayout({
+			...JSON.parse(inputValue || '{}'),
 		});
-	}
+	}, [inputValue]);
 
-	_handleOnDispatch(event) {
-		switch (event.type) {
-			case 'value':
-				this.dispatchEvent(event, 'fieldEdited', event.payload);
-				break;
-			case 'blur':
-				this.dispatchEvent(
-					event.payload,
-					'fieldBlurred',
-					event.payload.target.value
-				);
-				break;
-			case 'focus':
-				this.dispatchEvent(
-					event.payload,
-					'fieldFocused',
-					event.payload.target.value
-				);
-				break;
-			default:
-				console.error(new TypeError(`There is no type ${event.type}`));
-				break;
+	const handleClearClick = () => {
+		setLayout({});
+		onChange('');
+	};
+
+	const handleFieldChanged = event => {
+		const selectedItem = event.selectedItem;
+
+		if (selectedItem && selectedItem.layoutId) {
+			setLayout(selectedItem);
+			onChange(selectedItem);
 		}
-	}
-}
+	};
 
-LayoutSelector.STATE = {
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
+	const handleItemSelectorTriggerClick = event => {
+		event.preventDefault();
 
-	errorMessage: Config.string(),
+		const itemSelectorDialog = new ItemSelectorDialog({
+			eventName: `${portletNamespace}selectLayout`,
+			singleSelect: true,
+			title: Liferay.Language.get('page'),
+			url: itemSelectorURL,
+		});
 
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?bool}
-	 */
+		itemSelectorDialog.on('selectedItemChange', handleFieldChanged);
 
-	evaluable: Config.bool().value(false),
+		itemSelectorDialog.open();
+	};
 
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
+	return (
+		<ClayForm.Group style={{marginBottom: '0.5rem'}}>
+			<ClayInput.Group>
+				<ClayInput.GroupItem className="d-none d-sm-block" prepend>
+					<input name={name} type="hidden" value={inputValue} />
 
-	fieldName: Config.string(),
+					<ClayInput
+						className="bg-light"
+						disabled={disabled}
+						onClick={handleItemSelectorTriggerClick}
+						readOnly
+						type="text"
+						value={layout.name || ''}
+					/>
+				</ClayInput.GroupItem>
 
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
+				<ClayInput.GroupItem append shrink>
+					<ClayButton
+						disabled={disabled}
+						displayType="secondary"
+						onClick={handleItemSelectorTriggerClick}
+						type="button"
+					>
+						{Liferay.Language.get('select')}
+					</ClayButton>
+				</ClayInput.GroupItem>
 
-	label: Config.string(),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	name: Config.string().required(),
-
-	/**
-	 * @default '000000'
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	predefinedValue: Config.string().value('000000'),
-
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?bool}
-	 */
-
-	readOnly: Config.bool().value(false),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(bool|undefined)}
-	 */
-
-	repeatable: Config.bool().value(false),
-
-	/**
-	 * @default false
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(bool|undefined)}
-	 */
-
-	required: Config.bool().value(false),
-
-	/**
-	 * @default true
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(bool|undefined)}
-	 */
-
-	showLabel: Config.bool().value(true),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	spritemap: Config.string(),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	tip: Config.string(),
-
-	/**
-	 * @default undefined
-	 * @instance
-	 * @memberof LayoutSelector
-	 * @type {?(string|undefined)}
-	 */
-
-	value: Config.string(),
+				{layout.layoutId && (
+					<ClayInput.GroupItem shrink>
+						<ClayButton
+							disabled={disabled}
+							displayType="secondary"
+							onClick={handleClearClick}
+							type="button"
+						>
+							{Liferay.Language.get('clear')}
+						</ClayButton>
+					</ClayInput.GroupItem>
+				)}
+			</ClayInput.Group>
+		</ClayForm.Group>
+	);
 };
 
-Soy.register(LayoutSelector, templates);
+const LayoutSelectorProxy = connectStore(
+	({
+		emit,
+		itemSelectorURL,
+		name,
+		portletNamespace,
+		predefinedValue,
+		readOnly,
+		value,
+		...otherProps
+	}) => (
+		<FieldBaseProxy {...otherProps} name={name} readOnly={readOnly}>
+			<LayoutSelector
+				disabled={readOnly}
+				inputValue={value && value !== '' ? value : predefinedValue}
+				itemSelectorURL={itemSelectorURL}
+				name={name}
+				onChange={value => emit('fieldEdited', {}, value)}
+				portletNamespace={portletNamespace}
+			/>
+		</FieldBaseProxy>
+	)
+);
 
-export {LayoutSelector};
-export default LayoutSelector;
+const ReactLayoutSelectorAdapter = getConnectedReactComponentAdapter(
+	LayoutSelectorProxy,
+	'link_to_layout'
+);
+
+export {ReactLayoutSelectorAdapter};
+export default ReactLayoutSelectorAdapter;
