--- modules/apps/layout/layout-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/LayoutSelectorAdapter.soy
@@ -1,8 +1,0 @@
-{namespace LayoutSelectorAdapter}
-
-/**
- * LayoutAdapter
- */
-{template .render}
-	<div key="ddm-field-type-layout-selector"></div>
-{/template}
