--- modules/apps/layout/layout-dynamic-data-mapping-form-field-type/src/main/resources/META-INF/resources/LayoutSelectorRegister.soy
@@ -1,5 +1,0 @@
-{namespace LayoutleSelectorRegister}
-
-{deltemplate PageRenderer.RegisterFieldType variant="'link_to_layout'"}
-	{call LayoutSelector.render data="all" /}
-{/deltemplate}
